# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /satcat:
    get:
      operationId: getSatcat
      tags: [satcat]
      summary: List satellite catalog entries
      description: Paginated, filterable, sortable list of McDowell GCAT satellite catalog entries
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: search
          in: query
          schema:
            type: string
          description: Text search across name fields
        - name: owner
          in: query
          schema:
            type: string
          description: Filter by owner/country code
        - name: objectClass
          in: query
          schema:
            type: string
          description: "Filter by object class: P (payload), R (rocket body), D (debris), U (unknown)"
        - name: orbit
          in: query
          schema:
            type: string
          description: Filter by operational orbit type (LEO, MEO, GEO, HEO, etc.)
        - name: satState
          in: query
          schema:
            type: string
          description: Filter by satellite state
        - name: gunterType
          in: query
          schema:
            type: string
          description: Filter by Gunter's Space Page "Type / Application" classification
        - name: massMin
          in: query
          schema:
            type: number
          description: Minimum mass in kg (inclusive); entries with unknown mass are excluded
        - name: massMax
          in: query
          schema:
            type: number
          description: Maximum mass in kg (inclusive); entries with unknown mass are excluded
        - name: sort
          in: query
          schema:
            type: string
          description: Field to sort by
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: asc
      responses:
        "200":
          description: Paginated satellite catalog list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SatcatListResponse"
...
  /rpod/events:
    get:
      operationId: getRpodEvents
      tags: [rpod]
      summary: List RPOD events
      description: Flagged rendezvous/proximity events, paginated and sortable; each event carries all involved spacecraft
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [active, stale]
          description: Filter by event status
        - name: sort
          in: query
          schema:
            type: string
            enum: [tca, minRangeKm, relVelKmS, memberCount]
          description: Sort field (default tca)
        - name: order
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated RPOD event list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RpodEventList"
  /rpod/events/{id}:
    get:
      operationId: getRpodEvent
      tags: [rpod]
      summary: RPOD event detail
      description: Event with per-member catalog info and latest element sets for 3D plotting
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Event detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RpodEventDetail"
        "404":
          description: Event not found
...
components:
  schemas:
    RpodMember:
      type: object
      properties:
        norad:
          type: integer
        name:
          type: string
          nullable: true
        owner:
          type: string
          nullable: true
        state:
          type: string
          nullable: true
        objectClass:
          type: string
          nullable: true
        opOrbit:
          type: string
          nullable: true
        ldate:
          type: string
          nullable: true
        minRangeKm:
          type: number
          nullable: true
          description: Tightest pairwise range this member reached vs any other member (km)
        relVelKmS:
          type: number
          nullable: true
      required: [norad]
    RpodTle:
      type: object
      properties:
        line1:
          type: string
        line2:
          type: string
        epoch:
          type: string
        incDeg:
          type: number
        raanDeg:
          type: number
        eccentricity:
          type: number
        argPerigeeDeg:
          type: number
        meanAnomalyDeg:
          type: number
        meanMotionRevPerDay:
          type: number
      required: [line1, line2, epoch, incDeg, raanDeg, eccentricity, argPerigeeDeg, meanAnomalyDeg, meanMotionRevPerDay]
    RpodMemberDetail:
      allOf:
        - $ref: "#/components/schemas/RpodMember"
        - type: object
          properties:
            tle:
              $ref: "#/components/schemas/RpodTle"
              nullable: true
    RpodEvent:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        windowStart:
          type: string
        windowEnd:
          type: string
        tca:
          type: string
          description: Predicted time of closest approach (ISO)
        minRangeKm:
          type: number
        relVelKmS:
          type: number
        memberCount:
          type: integer
        widenedScan:
          type: boolean
          description: True when the cluster hit the member cap and a widened neighborhood scan ran
        firstDetectedAt:
          type: string
        updatedAt:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/RpodMember"
      required: [id, status, windowStart, windowEnd, tca, minRangeKm, relVelKmS, memberCount, widenedScan, firstDetectedAt, updatedAt, members]
    RpodEventList:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/RpodEvent"
        total:
          type: integer
        page:
          type: integer
        pages:
          type: integer
      required: [data, total, page, pages]
    RpodEventDetail:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
        windowStart:
          type: string
        windowEnd:
          type: string
        tca:
          type: string
        minRangeKm:
          type: number
        relVelKmS:
          type: number
        memberCount:
          type: integer
        widenedScan:
          type: boolean
        firstDetectedAt:
          type: string
        updatedAt:
          type: string
        members:
          type: array
          items:
            $ref: "#/components/schemas/RpodMemberDetail"
      required: [id, status, windowStart, windowEnd, tca, minRangeKm, relVelKmS, memberCount, widenedScan, firstDetectedAt, updatedAt, members]
...
    SatcatEntry:
      type: object
      properties:
        jcat:
          type: string
          description: Jonathan's catalog number
        satno:
          type: ["integer", "null"]
          description: US Space Command catalog number
        name:
          type: string
          description: Object name
        plName:
          type: ["string", "null"]
          description: Payload name
        ldate:
          type: ["string", "null"]
          description: Launch date (YYYY-MM-DD)
        lv:
          type: ["string", "null"]
          description: Launch vehicle
        lvFamily:
          type: ["string", "null"]
          description: Launch vehicle family
        site:
          type: ["string", "null"]
          description: Launch site code
        owner:
          type: ["string", "null"]
          description: Owner country or organization
        state:
          type: ["string", "null"]
          description: Country/state code
        objectClass:
          type: ["string", "null"]
          description: "Object class: P=payload, R=rocket body, D=debris, U=unknown"
        objType:
          type: ["string", "null"]
          description: More specific object type
        opOrbit:
          type: ["string", "null"]
          description: Operational orbit type (LEO, MEO, GEO, HEO, etc.)
        satState:
          type: ["string", "null"]
          description: Satellite state (Operational, Dead, Reentry, etc.)
        massKg:
          type: ["number", "null"]
          description: Dry mass in kilograms
        massEstimated:
          type: boolean
          description: True when massKg is a Bureau estimate rather than GCAT-catalogued data
        apogeeKm:
          type: ["number", "null"]
          description: Apogee altitude in km
        perigeeKm:
          type: ["number", "null"]
          description: Perigee altitude in km
        incDeg:
          type: ["number", "null"]
          description: Orbital inclination in degrees
        periodMin:
          type: ["number", "null"]
          description: Orbital period in minutes
        decayDate:
          type: ["string", "null"]
          description: Decay or reentry date
        gunterType:
          type: ["string", "null"]
          description: Gunter's Space Page "Type / Application" classification
        gunterNation:
          type: ["string", "null"]
          description: Gunter's Space Page "Nation" fact
        gunterOperator:
          type: ["string", "null"]
          description: Gunter's Space Page "Operator" fact
        gunterContractors:
          type: ["string", "null"]
          description: Gunter's Space Page "Contractors" fact
        gunterUrl:
          type: ["string", "null"]
          description: Cross-link to the full dossier on space.skyrocket.de
        gunterTitle:
          type: ["string", "null"]
          description: Dossier page title, for Krebs-format citations
        gunterRetrievedAt:
          type: ["string", "null"]
          description: ISO timestamp the dossier was retrieved
      required:
        - jcat
        - name
...
    SatcatListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/SatcatEntry"
        total:
          type: integer
          description: Total matching records
        page:
          type: integer
        limit:
          type: integer
        pages:
          type: integer
          description: Total pages
        filteredMassKg:
          type: number
          description: Total confirmed mass (kg) of ALL entries matching the current filters (not just this page)
        filteredEstMassKg:
          type: number
          description: Total Bureau-estimated mass (kg) of ALL entries matching the current filters
      required:
        - data
        - total
        - page
        - limit
        - pages
        - filteredMassKg
        - filteredEstMassKg
...