# lib/api-zod/src/generated/types/rpodEvent.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RpodMember } from './rpodMember';

export interface RpodEvent {
  id: number;
  status: string;
  windowStart: string;
  windowEnd: string;
  /** Predicted time of closest approach (ISO) */
  tca: string;
  minRangeKm: number;
  relVelKmS: number;
  memberCount: number;
  /** True when the cluster hit the member cap and a widened neighborhood scan ran */
  widenedScan: boolean;
  firstDetectedAt: string;
  updatedAt: string;
  members: RpodMember[];
}