# lib/api-zod/src/generated/types/rpodEventDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RpodMemberDetail } from './rpodMemberDetail';

export interface RpodEventDetail {
  id: number;
  status: string;
  windowStart: string;
  windowEnd: string;
  tca: string;
  minRangeKm: number;
  relVelKmS: number;
  memberCount: number;
  widenedScan: boolean;
  firstDetectedAt: string;
  updatedAt: string;
  members: RpodMemberDetail[];
}