# lib/api-zod/src/generated/types/rpodEventList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { RpodEvent } from './rpodEvent';

export interface RpodEventList {
  data: RpodEvent[];
  total: number;
  page: number;
  pages: number;
}