# lib/api-zod/src/generated/types/rpodMember.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface RpodMember {
  norad: number;
  name?: string | null;
  owner?: string | null;
  state?: string | null;
  objectClass?: string | null;
  opOrbit?: string | null;
  ldate?: string | null;
  /** Tightest pairwise range this member reached vs any other member (km) */
  minRangeKm?: number | null;
  relVelKmS?: number | null;
}